+++ docs/md/releasenotes.md
@@ -19,6 +19,7 @@
 - Add `KotlinValueParameterFilter` to allow filtering easily when visiting value parameters.
 - Implement `ClassVisitor` in `KotlinMetadataInitializer` to allow easier initialization of Kotlin metadata.
 - Implement `ClassVisitor` in `KotlinMetadataWriter` to allow easier writing of Kotlin metadata.
+- Add error handler callback to `KotlinMetadataInitializer`.
 
 ### Bug fixes